+++ src/main/java/com/gxdxx/instagram/domain/follow/api/FollowController.java
@@ -1,14 +1,14 @@
-package com.gxdxx.instagram.controller;
+package com.gxdxx.instagram.domain.follow.api;
 
-import com.gxdxx.instagram.dto.request.MessageListRequest;
-import com.gxdxx.instagram.dto.request.MessageSendRequest;
+import com.gxdxx.instagram.dto.request.FollowCreateRequest;
 import com.gxdxx.instagram.dto.response.ErrorResponse;
-import com.gxdxx.instagram.dto.response.MessageResponse;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.exception.InvalidRequestException;
-import com.gxdxx.instagram.domain.message.application.MessageCreateService;
-import com.gxdxx.instagram.domain.message.application.MessageQueryService;
+import com.gxdxx.instagram.domain.follow.application.FollowCreateService;
+import com.gxdxx.instagram.domain.follow.application.FollowDeleteService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -22,47 +22,45 @@ import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
 
-@Tag(name = "messages", description = "메시지 API")
+@Tag(name = "follows", description = "팔로우 API")
 @RequiredArgsConstructor
-@RequestMapping("/api/messages")
+@RequestMapping("/api/follows")
 @RestController
-public class MessageController {
+public class FollowController {
 
-    private final MessageCreateService messageCreateService;
-    private final MessageQueryService messageQueryService;
+    private final FollowCreateService followCreateService;
+    private final FollowDeleteService followDeleteService;
 
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "메시지 전송 성공",
+            @ApiResponse(responseCode = "201", description = "팔로우 성공",
                     content = @Content(schema = @Schema(implementation = SuccessResponse.class))),
             @ApiResponse(responseCode = "400", description = "잘못된 요청",
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
-    @Operation(summary = "메시지 전송 메소드", description = "메시지 전송 메소드입니다.")
+    @Operation(summary = "팔로우 메소드", description = "팔로우 메소드입니다.")
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public SuccessResponse createMessage(
-            @RequestBody @Valid MessageSendRequest request,
+    public SuccessResponse createFollow(
+            @RequestBody @Valid FollowCreateRequest request,
             BindingResult bindingResult,
             Principal principal
     ) {
         validateRequest(bindingResult);
-        return messageCreateService.sendMessage(request, principal.getName());
+        return followCreateService.createFollow(request, principal.getName());
     }
 
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "메시지 조회 성공",
-                    content = @Content(schema = @Schema(implementation = MessageResponse.class))),
+            @ApiResponse(responseCode = "204", description = "팔로우 취소 성공",
+                    content = @Content(schema = @Schema(implementation = SuccessResponse.class))),
             @ApiResponse(responseCode = "400", description = "잘못된 요청",
                     content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
-    @Operation(summary = "메시지 조회 메소드", description = "메시지 조회 메소드입니다.")
-    @GetMapping(value = "/{chat-room-id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public MessageResponse findMessages(
-            @RequestBody @Valid MessageListRequest request,
-            BindingResult bindingResult,
+    @Operation(summary = "팔로우 취소 메소드", description = "팔로우 취소 메소드입니다.")
+    @DeleteMapping(value = "/{id}", consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public SuccessResponse deleteFollow(
+            @Parameter(name = "id", description = "게시글의 id", in = ParameterIn.PATH) @PathVariable("id") Long id,
             Principal principal
     ) {
-        validateRequest(bindingResult);
-        return messageQueryService.findMessages(request, principal.getName());
+        return followDeleteService.deleteFollow(id, principal.getName());
     }
 
     private void validateRequest(BindingResult bindingResult) {
